refactor: add user domain to user package
