+++ app/src/main/java/com/codesoom/assignment/user/application/UserEmailDuplicationException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.errors;
+package com.codesoom.assignment.user.application;
 
 public class UserEmailDuplicationException extends RuntimeException {
     public UserEmailDuplicationException(String email) {
