+++ app/src/main/java/com/codesoom/assignment/user/dto/UserResultData.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dto;
+package com.codesoom.assignment.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
