+++ app/src/test/java/com/codesoom/assignment/user/application/UserServiceTest.java
@@ -1,11 +1,12 @@
-package com.codesoom.assignment.application;
+package com.codesoom.assignment.user.application;
 
-import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.domain.UserRepository;
-import com.codesoom.assignment.dto.UserModificationData;
-import com.codesoom.assignment.dto.UserRegistrationData;
-import com.codesoom.assignment.errors.UserEmailDuplicationException;
-import com.codesoom.assignment.errors.UserNotFoundException;
+import com.codesoom.assignment.user.application.UserService;
+import com.codesoom.assignment.user.domain.User;
+import com.codesoom.assignment.user.domain.UserRepository;
+import com.codesoom.assignment.user.dto.UserModificationData;
+import com.codesoom.assignment.user.dto.UserRegistrationData;
+import com.codesoom.assignment.user.application.UserEmailDuplicationException;
+import com.codesoom.assignment.user.application.UserNotFoundException;
 import com.github.dozermapper.core.DozerBeanMapperBuilder;
 import com.github.dozermapper.core.Mapper;
 import org.junit.jupiter.api.BeforeEach;
