+++ app/src/test/java/com/codesoom/assignment/user/controller/UserControllerTest.java
@@ -1,10 +1,11 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.user.controller;
 
-import com.codesoom.assignment.application.UserService;
-import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.dto.UserModificationData;
-import com.codesoom.assignment.dto.UserRegistrationData;
-import com.codesoom.assignment.errors.UserNotFoundException;
+import com.codesoom.assignment.user.application.UserService;
+import com.codesoom.assignment.user.controller.UserController;
+import com.codesoom.assignment.user.domain.User;
+import com.codesoom.assignment.user.dto.UserModificationData;
+import com.codesoom.assignment.user.dto.UserRegistrationData;
+import com.codesoom.assignment.user.application.UserNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
